LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.drawing.style.letterSpacing.measure-expected.txt

2d.text.drawing.style.letterSpacing.measure
Testing letter spacing with different length units
Actual output:

FAIL Testing letter spacing with different length units assert_equals: ctx.letterSpacing === '0px' (got [undefined], expected 0px[string]) expected (string) "0px" but got (undefined) undefined
